file upload error resolved

+++ src/utils/fileUpload.js
@@ -13,7 +13,9 @@ const uploadOnCloudinary = async (localFilePath) => {
     console.log("file is uploaded on cloudinary", response.url);
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); //removes the locally saved temporary file as the upload operation got failed
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath); //removes the locally saved temporary file as the upload operation got failed
+    }
     return null;
   }
 };
